Add T world-side runtime and the behavior end-to-end
Milestone 5's core: T scripts now drive real entity behavior on the
fixed tick, exercising the whole cook -> link -> VM -> host ->
component-write -> determinism stack.

- ScriptLink resolves a validated, world-agnostic module against a
  World: registers the module's script components (RegisterComponentRaw,
  before first entity), resolves tag names to GameplayTagIds, and
  resolves component/field paths to ComponentId plus byte offset plus
  scalar, from the module schema (script components) or an explicit
  native host-component table (Transform, whose Transform3f has no
  reflected recursion). Link failure names the missing symbol.
- WorldScriptHost is the real ScriptHostCallHandler: component field
  load/store through World::GetComponentRaw with f32<->f64 (and int/
  bool/entity) conversion between storage and the VM's 64-bit registers,
  GameplayTagContainer grant/revoke, and DeterministicRandom streams
  seeded from world seed, entity, and tick.
- ScriptBehaviorSystem invokes behavior callbacks on the fixed tick:
  spawn once (lazily), then the per-state or bare fixed callback, with
  structural changes deferred to a CommandBuffer flushed at the tick
  boundary and state transitions applied after each callback.
- The compiler's Transform fields become f32 (matching native storage)
  so the host conversion is correct.
- Fitness ctests: check_script_layering (the VM core takes no upward
  gameplay/runtime include; bridges are the only seam) and
  check_script_host_surface (no wall-clock/IO/thread symbols in the
  runtime core). Both verified to catch seeded violations.

Tests: the pickup behavior cooked and linked into a World, ticked, and
its bob trajectory checked against a reference plus reproduced bit-for-
bit across two worlds; a link-error test. Full suite 1253 tests, 100%.

Deferred M5 (need native systems AbilityKit does not yet expose): the
ability bridge, trigger/interaction systems, and the full
ScriptComponentSerializer, recorded in the spec.

+++ engine/src/script/compiler/ScriptCodegen.cpp
@@ -2758,9 +2758,12 @@ namespace
 
 const ScriptHostDecls& DefaultScriptHostDecls()
 {
+    // Native LocalTransform stores Transform3f (f32 position/scale), so the
+    // script-visible fields are f32 and the field-bind scalar is Float; the
+    // world-side host converts f32 storage to the VM's f64 registers.
     static const ScriptHostFieldDecl kTransformFields[] = {
-        {"position", "local.position", ScriptScalarKind::Double, 3},
-        {"scale", "local.scale", ScriptScalarKind::Double, 3},
+        {"position", "local.position", ScriptScalarKind::Float, 3},
+        {"scale", "local.scale", ScriptScalarKind::Float, 3},
     };
     static const ScriptHostComponentDecl kComponents[] = {
         {"Transform", kTransformFields},
